% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_funs.R
\name{trace_funs}
\alias{trace_funs}
\alias{untrace_funs}
\title{Trace or untrace ggplot functions}
\usage{
trace_funs(
  pkgs = "ggplot2",
  prefixes = c("annotate_", "coord_", "facet_", "geom_", "scale_", "stat_", "theme"),
  exceptions = c("scale_type", "theme_get", "coord_munch", "geom_blank")
)

untrace_funs(
  pkgs = "ggplot2",
  prefixes = c("annotate_", "coord_", "facet_", "geom_", "scale_", "stat_", "theme"),
  exceptions = c("scale_type", "theme_get", "coord_munch")
)
}
\arguments{
\item{pkgs}{packages where to find functions to trace/untrace}

\item{prefixes}{prefixes of those functions}

\item{exceptions}{exceptions}
}
\description{
Trace or untrace ggplot functions
}
